fix: fix all compilation errors found in CI
Fix all compilation errors found in CI build:

1. Windows Sys Import Error (E0432)
   File: src/sandbox.rs:420
   Error: unresolved import windows_sys::Win32::System::ProcessStatus::GetProcessTimes
   Fix: Remove unused import GetProcessTimes
   Change: Only keep OpenProcess import

2. Unused Import Warning
   File: src/ipc.rs:6
   Warning: unused import self
   Fix: Remove self from ring::signature imports
   Change: use ring::signature::{Ed25519KeyPair, KeyPair, UnparsedPublicKey, ED25519}

3. Type Mismatch Error (E0308)
   File: src/ipc.rs:270
   Error: expected Ed25519KeyPair, found Result<Ed25519KeyPair, KeyRejected>
   Fix: Handle Result type from from_seed_unchecked
   Change: Add ? operator to propagate error

4. Type Mismatch Error (E0308)
   File: src/sandbox.rs:472
   Error: expected u64, found usize
   Fix: Cast result to u64
   Change: Ok((memory_info.WorkingSetSize / 1024 / 1024) as u64)

These fixes resolve all compilation errors found during cargo check.

Modified files:
- src/sandbox.rs: fix Windows sys import and type casting
- src/ipc.rs: remove unused import and handle Result type

+++ src/sandbox.rs
@@ -417,7 +417,6 @@ impl Sandbox {
     /// 获取 CPU 使用率（Windows 实现）
     #[cfg(target_os = "windows")]
     fn get_cpu_usage(&self, pid: u32) -> Result<f64, String> {
-        use windows_sys::Win32::System::ProcessStatus::GetProcessTimes;
         use windows_sys::Win32::System::Threading::OpenProcess;
 
         unsafe {
@@ -470,7 +469,7 @@ impl Sandbox {
                 return Err("无法获取进程内存信息".to_string());
             }
 
-            Ok(memory_info.WorkingSetSize / 1024 / 1024)
+            Ok((memory_info.WorkingSetSize / 1024 / 1024) as u64)
         }
     }
